userAPI/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
psycopg2-binary
